# public/api/openapi.yaml
openapi: 3.1.0
info:
  title: J1.Notes API
  description: |
    REST API for **J1.Notes** — a self-hosted, privacy-focused notes application.

    ## Authentication
    When app-lock is enabled, all API requests require a valid `j1notes-auth` session cookie
    obtained via `POST /api/auth/login`. Routes marked as **public** are always accessible.

    ## Base URL
    `http://localhost:3000` (default Docker setup)
  version: "1.0.3"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: J1.Notes on GitHub
    url: https://github.com/x3kim/J1.Notes
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/notes/{id}:
    parameters:
      - $ref: "#/components/parameters/NoteId"

    get:
      tags: [Notes]
      summary: Get a single note
      responses:
        "200":
          description: Note object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Note not found

    patch:
      tags: [Notes]
      summary: Update a note
      description: Partially updates a note. Only the provided fields are changed. Version history is created automatically for substantial content changes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "500":
          description: Server error

    delete:
      tags: [Notes]
      summary: Trash or permanently delete a note
      parameters:
        - name: hard
          in: query
          description: If `true`, permanently deletes the note. Otherwise moves it to trash.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Note deleted / moved to trash
        "500":
          description: Server error
...
  /api/labels:
    get:
      tags: [Labels]
      summary: List all labels
      responses:
        "200":
          description: Array of labels sorted alphabetically
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"

    post:
      tags: [Labels]
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: Work
                color:
                  type: string
                  nullable: true
                  example: "#facc15"
      responses:
        "201":
          description: Label created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "400":
          description: Label already exists or invalid data
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in (public)
      description: Validates the password/PIN and sets the `j1notes-auth` session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  description: Password or PIN depending on auth_type
      responses:
        "200":
          description: Login successful — session cookie is set
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          description: Auth not enabled
        "401":
          description: Wrong password / PIN
        "500":
          description: Server error
...
  /api/avatar:
    post:
      tags: [Profile]
      summary: Upload an avatar image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PNG, JPG or WebP — max 2 MB
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    example: /uploads/avatar-1714000000000.jpg
        "400":
          description: No file provided or unsupported type
        "413":
          description: File too large

    delete:
      tags: [Profile]
      summary: Delete avatar file
      parameters:
        - name: file
          in: query
          required: true
          schema:
            type: string
          description: URL of the avatar to delete (e.g. /uploads/avatar-123.jpg)
      responses:
        "200":
          description: Deleted (or gracefully ignored)

  # ── Upload ─────────────────────────────────────────────────────────────────
...
components:
  parameters:
    NoteId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Note UUID
  schemas:
    Note:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        content_text:
          type: string
          nullable: true
          description: HTML content (from rich-text editor)
        color:
          type: string
          nullable: true
          description: CSS color string (e.g. #facc15)
        bg_image:
          type: string
          nullable: true
        pinned:
          type: boolean
        archived:
          type: boolean
        position:
          type: integer
        reminder_at:
          type: string
          format: date-time
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        checklist_items:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
...
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content_text:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        bg_image:
          type: string
          nullable: true
        pinned:
          type: boolean
        archived:
          type: boolean
        position:
          type: integer
        reminder_at:
          type: string
          format: date-time
          nullable: true
        label_ids:
          type: array
          items:
            type: string
        checklist_items:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              checked:
                type: boolean
...
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        checked:
          type: boolean
        position:
          type: integer
    Attachment:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
    Label:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
...